requirements.txt
requests
loguru
peewee
pytesseract
PyAutoGUI
selenium
python-dotenv